Fix initial timestamp
The server timestamps of the data (read) or newData (write/patch) snapshot do not necessary match the rulesets “now” timestamp.

The tests are unstable, relying on the snapshot and ruleset state `now` variable being created in the same nanosecond.

This fix attaches the timestamp to the snapshot and uses it to evaluate the ruleset.

API changes (backward compatible):
- `targaryen/lib/rule-data-snapshot.RuleDataSnapshot(data[, now], path)`;
- `targaryen/lib/rule-data-snapshot.RuleDataSnapshot.convert(data[, now])`;
- `targaryen/lib/helpers.makeNewDataSnap(path, data[, now])`;
- `targaryen.setFirebaseData(data[, now])`;
- `targayen/lib/ruleset.prototype.tryRead(path, root, auth[, now])`;
- `targayen/lib/ruleset.prototype.tryWrite(path, root, newData, auth[, skipWrite[, skipValidate[, skipOnNoValue[, now]]]])`;
- `targayen/lib/ruleset.prototype.tryPatch(path, root, newData, auth[, skipWrite[, skipValidate[, skipOnNoValue[, now]]]])`;
- jasmine matchers gain an optional “now” argument;
- chai `write` and `patch` methods gain an optional “now” argument.

`now` default to the value returned by`Date.now()` if not provided.

Fix #29.